+++ lib/common/service/theme/theme_service.dart
@@ -1,13 +1,28 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_template/common/service/theme/i_theme_service.dart';
 
 /// A service that stores and retrieves app theme mode.
-abstract class IThemeService extends ChangeNotifier {
-  /// Current app theme mode.
-  ThemeMode get currentThemeMode;
+class ThemeService extends IThemeService {
+  late ThemeMode _themeMode;
 
-  /// Switch theme mode between light and dark.
-  void switchTheme();
+  @override
+  ThemeMode get currentThemeMode => _themeMode;
 
-  /// Select theme mode.
-  void updateThemeMode(ThemeMode newThemeMode);
+  /// Create an instance [ThemeService].
+  ThemeService(this._themeMode);
+
+  @override
+  void switchTheme() {
+    _themeMode == ThemeMode.light ? _themeMode = ThemeMode.dark : _themeMode = ThemeMode.light;
+    notifyListeners();
+  }
+
+  @override
+  void updateThemeMode(ThemeMode newThemeMode) {
+    if (newThemeMode == _themeMode) return;
+
+    _themeMode = newThemeMode;
+
+    notifyListeners();
+  }
 }
